Issue #60 fix regression for missing items in the user menu.
The missing menu items were caused with issue #47.

Since we now clean up the duplicate envbars in $CFG->additionalhtml
when the settings page is saved. It does not matter that it shows up in
the generic search results.

The renderer was previously creating the labels to the other sites based
on the url that was available in the link. This will restore that link
and label.